Move CacheBustCondition to its own enums file
refs TS-272
