+++ django/thunderstore/frontend/templatetags/dynamic_html.py
@@ -1,7 +1,8 @@
 from django import template
 from django.utils.safestring import mark_safe
 
-from thunderstore.cache.cache import CacheBustCondition, cache_function_result
+from thunderstore.cache.cache import cache_function_result
+from thunderstore.cache.enums import CacheBustCondition
 from thunderstore.frontend.models import DynamicHTML
 
 register = template.Library()
