+++ django/thunderstore/repository/api/experimental/views/package.py
@@ -4,7 +4,8 @@ from rest_framework.exceptions import ValidationError
 from rest_framework.generics import ListAPIView, RetrieveAPIView, get_object_or_404
 from rest_framework.pagination import CursorPagination
 
-from thunderstore.cache.cache import CacheBustCondition, ManualCacheCommunityMixin
+from thunderstore.cache.cache import ManualCacheCommunityMixin
+from thunderstore.cache.enums import CacheBustCondition
 from thunderstore.repository.api.experimental.serializers import (
     PackageSerializerExperimental,
 )
