+++ django/thunderstore/repository/api/experimental/views/package_version.py
@@ -2,7 +2,8 @@ from drf_yasg.utils import swagger_auto_schema
 from rest_framework.exceptions import ValidationError
 from rest_framework.generics import RetrieveAPIView, get_object_or_404
 
-from thunderstore.cache.cache import CacheBustCondition, ManualCacheCommunityMixin
+from thunderstore.cache.cache import ManualCacheCommunityMixin
+from thunderstore.cache.enums import CacheBustCondition
 from thunderstore.repository.api.experimental.serializers import (
     PackageVersionSerializerExperimental,
 )
